update from float to double for calculations
We now use double instead of float for all important values that will be used for calculations.
- update from float to double type
- update string formatting in Vector3 class (now more accurate, but not optimized for storage yet)

+++ SimpleStepWriterSolution/SimpleStepWriter/Helper/Matrix3x3.cs
@@ -5,28 +5,28 @@ namespace SimpleStepWriter.Helper
 {
     public class Matrix3x3
     {
-        float[,] matrix;
+        double[,] matrix;
         
-        public float this[int i, int j]
+        public double this[int i, int j]
         {
             get { return matrix[i, j]; }            
         }
 
         #region properties
-        public float A11 { get { return matrix[0, 0]; } set { matrix[0, 0] = value; } }
-        public float A12 { get { return matrix[0, 1]; } set { matrix[0, 1] = value; } }
-        public float A13 { get { return matrix[0, 2]; } set { matrix[0, 2] = value; } }
-        public float A21 { get { return matrix[1, 0]; } set { matrix[1, 0] = value; } }
-        public float A22 { get { return matrix[1, 1]; } set { matrix[1, 1] = value; } }
-        public float A23 { get { return matrix[1, 2]; } set { matrix[1, 2] = value; } }
-        public float A31 { get { return matrix[2, 0]; } set { matrix[2, 0] = value; } }
-        public float A32 { get { return matrix[2, 1]; } set { matrix[2, 1] = value; } }
-        public float A33 { get { return matrix[2, 2]; } set { matrix[2, 2] = value; } }
+        public double A11 { get { return matrix[0, 0]; } set { matrix[0, 0] = value; } }
+        public double A12 { get { return matrix[0, 1]; } set { matrix[0, 1] = value; } }
+        public double A13 { get { return matrix[0, 2]; } set { matrix[0, 2] = value; } }
+        public double A21 { get { return matrix[1, 0]; } set { matrix[1, 0] = value; } }
+        public double A22 { get { return matrix[1, 1]; } set { matrix[1, 1] = value; } }
+        public double A23 { get { return matrix[1, 2]; } set { matrix[1, 2] = value; } }
+        public double A31 { get { return matrix[2, 0]; } set { matrix[2, 0] = value; } }
+        public double A32 { get { return matrix[2, 1]; } set { matrix[2, 1] = value; } }
+        public double A33 { get { return matrix[2, 2]; } set { matrix[2, 2] = value; } }
         #endregion        
 
         public Matrix3x3()
         {
-            matrix = new float[3, 3] {
+            matrix = new double[3, 3] {
                 { 0, 0, 0 },
                 { 0, 0, 0 },
                 { 0, 0, 0 }
@@ -35,18 +35,18 @@ namespace SimpleStepWriter.Helper
 
         public Matrix3x3
         (
-            float a11,
-            float a12,
-            float a13,
-            float a21,
-            float a22,
-            float a23,
-            float a31,
-            float a32,
-            float a33
+            double a11,
+            double a12,
+            double a13,
+            double a21,
+            double a22,
+            double a23,
+            double a31,
+            double a32,
+            double a33
         )
         {
-            matrix = new float[3, 3] {
+            matrix = new double[3, 3] {
                 { a11, a12, a13 },
                 { a21, a22, a23 },
                 { a31, a32, a33 }
@@ -67,7 +67,7 @@ namespace SimpleStepWriter.Helper
         {
             var x = obj as Matrix3x3;
             return x != null &&
-                   EqualityComparer<float[,]>.Default.Equals(matrix, x.matrix) &&
+                   EqualityComparer<double[,]>.Default.Equals(matrix, x.matrix) &&
                    A11 == x.A11 &&
                    A12 == x.A12 &&
                    A13 == x.A13 &&
@@ -82,7 +82,7 @@ namespace SimpleStepWriter.Helper
         public override int GetHashCode()
         {
             var hashCode = -929212920;
-            hashCode = hashCode * -1521134295 + EqualityComparer<float[,]>.Default.GetHashCode(matrix);
+            hashCode = hashCode * -1521134295 + EqualityComparer<double[,]>.Default.GetHashCode(matrix);
             hashCode = hashCode * -1521134295 + A11.GetHashCode();
             hashCode = hashCode * -1521134295 + A12.GetHashCode();
             hashCode = hashCode * -1521134295 + A13.GetHashCode();
@@ -132,25 +132,25 @@ namespace SimpleStepWriter.Helper
         {
             // calc radian
             var deg = new Vector3();
-            deg.X = (float)Math.PI * euler.X / 180.0f;
-            deg.Y = (float)Math.PI * euler.Y / 180.0f;
-            deg.Z = (float)Math.PI * euler.Z / 180.0f;
+            deg.X = Math.PI * euler.X / 180.0d;
+            deg.Y = Math.PI * euler.Y / 180.0d;
+            deg.Z = Math.PI * euler.Z / 180.0d;
 
             // return rotation matrix
             return new Matrix3x3
                     (
                         // first row
-                        (float)(Math.Cos(deg.Y) * Math.Cos(deg.Z) + Math.Sin(deg.Y) * Math.Sin(deg.X) * Math.Sin(deg.Z)),
-                        (float)(Math.Cos(deg.Z) * Math.Sin(deg.Y) * Math.Sin(deg.X) - Math.Sin(deg.Z) * Math.Cos(deg.Y)),
-                        (float)(Math.Cos(deg.X) * Math.Sin(deg.Y)),
+                        (Math.Cos(deg.Y) * Math.Cos(deg.Z) + Math.Sin(deg.Y) * Math.Sin(deg.X) * Math.Sin(deg.Z)),
+                        (Math.Cos(deg.Z) * Math.Sin(deg.Y) * Math.Sin(deg.X) - Math.Sin(deg.Z) * Math.Cos(deg.Y)),
+                        (Math.Cos(deg.X) * Math.Sin(deg.Y)),
                         // second row
-                        (float)(Math.Cos(deg.X) * Math.Sin(deg.Z)),
-                        (float)(Math.Cos(deg.Z) * Math.Cos(deg.X)),
-                        (float)(-Math.Sin(deg.X)),
+                        (Math.Cos(deg.X) * Math.Sin(deg.Z)),
+                        (Math.Cos(deg.Z) * Math.Cos(deg.X)),
+                        (-Math.Sin(deg.X)),
                         // third row
-                        (float)(Math.Sin(deg.Z) * Math.Cos(deg.Y) * Math.Sin(deg.X) - Math.Sin(deg.Y) * Math.Cos(deg.Z)),
-                        (float)(Math.Sin(deg.Y) * Math.Sin(deg.Z) + Math.Cos(deg.Z) * Math.Cos(deg.Y) * Math.Sin(deg.X)),
-                        (float)(Math.Cos(deg.X) * Math.Cos(deg.Y))
+                        (Math.Sin(deg.Z) * Math.Cos(deg.Y) * Math.Sin(deg.X) - Math.Sin(deg.Y) * Math.Cos(deg.Z)),
+                        (Math.Sin(deg.Y) * Math.Sin(deg.Z) + Math.Cos(deg.Z) * Math.Cos(deg.Y) * Math.Sin(deg.X)),
+                        (Math.Cos(deg.X) * Math.Cos(deg.Y))
                    );
         }
